rename products to products-schema
