+++ src/server/db/schema/index.ts
@@ -5,7 +5,7 @@ import { expenseCategory } from "./expense-category";
 import { incomes } from "./incomes-schema";
 import { sales } from "./sales-schema";
 import { users, accounts, sessions, verificationTokens } from "./auth-schema";
-import { products } from "./products";
+import { products } from "./products-schema";
 import { cashRegisterItem } from "./cash-register-items";
 import { customers } from "./customers";
 import { incomeItem } from "./income-items";
