+++ src/server/db/schema/products-schema.ts
@@ -1,6 +1,6 @@
 import { sql } from "drizzle-orm";
 import { createTable } from "./table-creator";
-import { serial, text, decimal, integer, timestamp } from "drizzle-orm/pg-core";
+import { serial, text, decimal, timestamp } from "drizzle-orm/pg-core";
 
 /**
  * Products kept in stock for sales.
@@ -10,7 +10,7 @@ export const products = createTable("products", {
   name: text("name").notNull(),
   cost: decimal("cost", { precision: 15, scale: 2 }).notNull(),
   price: decimal("price", { precision: 15, scale: 2 }).notNull(),
-  quantity: integer("quantity").notNull().default(0),
+  quantity: decimal("quantity").notNull().default("0"),
   createdAt: timestamp("createdAt", { withTimezone: true })
     .default(sql`CURRENT_TIMESTAMP`)
     .notNull(),
