+++ src/server/queries/stock-queries.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { db } from "../db";
-import { products } from "../db/schema/products";
+import { products } from "../db/schema/products-schema";
 import { eq } from "drizzle-orm";
 
 // Delete a stock item (product) by ID
